Part 2 of breaking distributor -> querier package coupling: move chunk metrics to `pkg/querier/stats`
